CRAYSAT-1839: Fix kubernetes version to match version in CSM
The version of the Python kubernetes client package was ahead of the
version of Kubernetes running in CSM 1.5, which resulted in some
compatibility issues. For example, `CoreV1Api.list_node` did not work if
a condition with a type of "ReadonlyFilesystem" was present on any of
the Kubernetes nodes in the cluster.

See https://github.com/kubernetes-client/python?tab=readme-ov-file#compatibility
for details on the compatibility of the Python `kubernetes` package with
the Kubernetes cluster version.

Also update the version of csm-api-client, which has been updated to
relax the kubernetes version constraint and allow versions >= 22.0
rather than only >= 23.0.

Test Description:
Built the container image, downloaded to odin, and ran the
`CoreV1Api.list_node` method in a Python interpreter in the `sat bash`
shell.

Also ran `sat bootprep run` against a simple input file that defined a
single CFS configuration, built an image from a recipe, and then
configured that image with CFS. This was just to ensure that nothing
using the K8s API in that code path was impacted.