CRAYSAT-1866: Fix `sat status` to only make necessary API requests
Fix the `sat status` command to only make the necessary API requests to
CSM services. Specifically, the `BOSStatusModule`, `SLSStatusModule`,
and the `CFSStatusModule`, only apply to components of type `Node`, and
so they should not be used when getting status of components of other
types, e.g. `NodeBMC`. Using those modules in that case results in many
unneccessary API requests to BOS, SLS, and CFS.

Simplify code a little by making `component_types` a required argument
for the `get_relevant_modules` and `get_poopulated_rows` methods. It was
confusing to have a lot of code that handled `None` when it really would
never be `None` because the corresponding argparse argument had a
default value of `["Node"]`.

Test Description:
Unit tests pass. Tested by running `sat status` against drax with
debug logging enabled and different values for `--types`. E.g. tested with `--types
NodeBMC RouterBMC` and verified that no requests were made to BOS, CFS,
or SLS. Then tested with `--types Node` (the default) and `--types Node
NodeBMC` to verify that requests were made to BOS, CFS, and SLS.