CRAYSAT-1100:Improve sat bootsys platform-services
IM:CRAYSAT-1100
Reviewer:Ryan

Implementation to improve `sat bootsys shutdown --stage platform-services`.
If containers fail to stop, automate the procedure of trying to stop
them again.

Improve the unit tests for ContainerStopProcess. The main improvement is
that we change the way that the unit tests verify whether the process
was successful by checking the values added to the `results_queue`
instead of looking at the `success` attribute, which will not be visible
to the parent process.

Additionally, change thread to process where appropriate in naming of
variables. E.g., change `cst` to `csp` for `ContainerStopProcess`
instead of `ContainerStopThread`.

Test Description:
Unit tests pass.