CRAYSAT-1100: Improve ContainerStopProcess unit tests
Improve the unit tests for ContainerStopProcess. The main improvement is
that we change the way that the unit tests verify whether the process
was successful by checking the values added to the `results_queue`
instead of looking at the `success` attribute, which will not be visible
to the parent process.

Additionally, change thread to process where appropriate in naming of
variables. E.g., change `cst` to `csp` for `ContainerStopProcess`
instead of `ContainerStopThread`.

Test Description:
Unit tests pass.