Addressing the value rounding bug
IM:CRAYSAT-1869
Reviewer:Ryan
There is a rounding error when SAT reports status on percentage done. It rounds up
to 100% when the actual number reported by BOS is something like 99.9475.For
anything other than 100%, most people would not care. But 100% means it is all done,
so why would SAT still be waiting for it to go over 100%

Hence, Updating the rounding to 2 decimal places instead of no decimal places and then
not rounding if the value above 99 to address the bug and showing upto 10 decimal places.

The same case is added part of the unit test for the class.