CRAYSAT-1861: Add unit tests for cron job enable/disable
Restructure the code for cron job enable/disable to make it easier to
unit test and then add unit tests. These unit tests actually require a
version of `sed` compatible with the version we expect on the management
node environment, but it felt important enough to test that the sed
command actually does what it's supposed to do.