Updating acceptable checks in ceph helath validation stage
IM:CRAYSAT-1581
Reviewer:Ryan

During the ceph health validation,in ncn-power stage.PG_NOT_DEEP_SCRUBBED
check would be in health_warn status and that would abort the next steps following
ceph health check as pgs not deep-scrubbed in time. Since it can be ignored, we are
adding it to the acceptable checkswhich can be not considered while returning the ceph health.