CRAYSAT-1861: Unmount Ceph/s3fs on ncn-m001 during shutdown
Modify the `sat bootsys shutdown --stage ncn-power` command to
do the following on ncn-m001:

* Find all mounts of filesystems of type `ceph` or `fuse.s3fs`
* Find all mounts of filesystems on RBD devices
* For each mount point identified, check whether there is any activity
  on that mount point that would prevent it being unmounted. Prompt the
  admin when they have stopped the processes before continuing.
* Disable systemd cron job that ensures Ceph and s3fs mounts remain
  mounted.
* Unmount each mount point
* Unmap all RBD devices

Modify the `sat bootsys boot --stage ncn-power` command to re-enable the
systemd cron job that ensures Ceph/s3fs mounts are mounted on ncn-m001.

Prior to this change, only the unmapping of RBD devices was performed.
We want to ensure that all usages of Ceph on ncn-m001 are gracefully
stopped before shutting down the storage nodes so that ncn-m001 will
shut down gracefully.

Test Description:
Added many unit tests.

Testing is needed on a real system as part of a full-system shutdown.