CRAYSAT-1861: Ensure host keys are filtered in `sat bootsys`
The `FilteredHostKeys` class was created to work around an issue with
Paramiko's loading of SSH host keys that was causing a serious
performance problem when the known_hosts file was large. However, this
class was not being used to filter host keys in all locations where a
Paramiko `SSHClient` was being created. Update the locations which were
not filtering host keys to do so.

Test Description:
Needs to be tested as part of a full-system power off/on on a real
system.