CRAYSAT-1859: Improve BOS failure handling in `sat bootsys`
If `sat bootsys` fails to create a BOS session for a BOS session
template, it will still attempt to check the status of a non-existent
BOS session, which results in a `TypeError` traceback when attempting to
get information about a BOS session with `None` as its ID.

Test Description:
Tested on rocket by passing in a non-existent BOS session template to
`sat bootsys boot --stage bos-operations`. Tested with and without this
fix to see the difference. With this fix, a traceback no longer occurs.

Tested also with a valid BOS session template to ensure it still works
as it should, waiting on the BOS session which is created.