CRAYSAT-1878: Remove automatic cronjob recreation from `bootsys`
Remove the step that automatically checks for and re-creates stuck
Kubernetes CronJobs from the `platform-services` stage of `sat bootsys
boot`. This should not be necessary anymore starting in Kubernetes 1.21,
which made a new CronJobControllerV2 the default.

In addition, improve the logic of the HMSDiscoveryScheduledWaiter, so
that it will more reliably detect when an `hms-discovery` Job has been
scheduled for the CronJob. Pass in an explicit `start_time`, so that we
can look for any jobs created for the CronJob after it is re-enabled.
This ensures we won't miss the first one, which could be scheduled
between when we set `suspend=False` on the CronJob and when we create
the `HMSDiscoveryScheduledWaiter`.

Test Description:
Tested on rocket as follows:

* Suspended the `hms-discovery` CronJob
* Ran `sat bootsys boot --stage cabinet-power`
* Verified that it correctly identified when the CronJob was scheduled