CRAYSAT-1890: Fix `sat.waiting` logger
The `LOGGER` was incorrectly created with `__file__` as the logger name
instead of `__name__`. As a result, it was not included under the `sat`
package logger's hierarchy, and the expected log handlers were not
applied to messages logged by this module. Fix this by specifying
`__name__` instead, so the logger has the name `sat.waiting`, which is
under the `sat` logger.

Test Description:
Ran `sat bootsys shutdown --stage capture-state` followed by `sat
bootsys boot --stage k8s-check --k8s-timeout 10` to force a
`Waiter` subclass to timeout, so I could observe the error log message.
As expected, it timed out waiting on new pods that were created by K8s
Jobs to reach a running or completed state, and it emitted the correctly
formatted log message of the form:

```
ERROR: Waiting for condition "..." timed out after 10 seconds
```