CRAYSAT-1917: Add unit tests for `jinja_rendered` decorator
Add unit tests for usage of the `jinja_rendered` decorator on properties
of the `BaseInputItem`.

Modify these unit tests to simplify them slightly and to ensure that the
`SomeTestItem` class calls the `BaseInputItem` constructor, so that it
sets the `jinja_env` instance attribute, which is used by the
`jinja_rendered` decorator.

Test Description:
Unit tests pass.