CRAYSAT-1896: Fix handling of "latest" product-based layers
If the value "latest" is specified for the version of a product in a
product-based layer in a CFS configuration in a bootprep input file,
pass in `None` to `from_product_catalog` to properly get the product
catalog data associated with the latest version of the product.

Test Description:
Not tested yet.