CRAYSAT-1896: Add `--cfs-version` option to `sat bootprep run`
Add the `--cfs-version` command-line option to `sat bootprep run`. This
option is used to select the version of the CFS API to use when creating
CFS configurations or checking for the existence of CFS configurations
when creating other items in the bootprep input file.

Add this option to the man page as well.

Test Description:
Built man page and examined in `man`.

Other testing yet to be performed on a system.