CRAYSAT-1896: Validate playbook in validation method instead of schema
Instead of requiring that the playbook be present in the bootprep input
file through the schema, validate it specifically when `sat` is
configured to use the CFS v3 API, since the CFS v2 API does not require
that CFS configuration layers have a playbook. This will allow people to
continue using old bootprep input files that have CFS configuration
layers that do not specify a playbook as long as they use CFS v2. This
is slightly less disruptive than making a schema change.

Test Description:
Unit tests only so far.