CRAYSAT-1896: Fix validation of the `InputConfiguration`
The `InputConfigurationLayer` instances stored in the `layers` attribute
of the `InputConfiguration` class were not being validated when the
`validate` method of the `InputConfiguration` was called because this
class did not have any validation methods. Add a validation method that
validates the layers, and make the `InputConfigurationLayer` class
inherit from `Validatable`, so that it has a `validate` method itself.

This has not been added for `AdditionalInventoryLayer` yet because there
are currently no validation methods on that class.

Note that an `index` is added to the constructor of the
`InputConfigurationLayer` class to allow for the `__str__` to
unambiguously refer to the layer within the CFS configuration.

Test Description:
Unit tests have been added to test the `validate` method on the
`InputConfiguration` class.