CRAYSAT-1896: Add Jinja template support to additional_inventory
Add support for Jinja template rendering to the `name`, `url`, `branch`,
and `commit` properties of the `additional_inventory` of a CFS
configuration specified in a bootprep input file.

This is useful for writing automated tests that obtain this information
from a system and pass it in through variables with `--vars-file`.

Test Description:
The added unit tests pass, and it will be tested on rocket as well.