CRAYSAT-1896: Fix `sat bootprep` to include False `ims_require_dkms`
Fixed `sat bootprep` to include the value of `ims_require_dkms` in the
created CFS configurations, even when explicitly set to `False`.

Test Description:
Tested on rocket with a `bootprep` input file that defines a CFS
configuration that contains two layers, one which sets the
`ims_require_dkms` property to `True` and one which sets it to `False`.
Verified that the resulting CFS configuration contained the value for
the property in both layers. Tested in dry-run mode and regular mode for
both CFS v2 and v3.