CRAYSAT-1948: Fix traceback in `sat swap blade`
Fix a traceback that occurs in `sat swap blade` when the node classes
reported by HSM includes both "Mountain" and "Hill" for a single blade.

This probably indicates some sort of configuration problem in HSM, but
the behavior of `sat swap blade` is the same for both Mountain and Hill,
so we can log a warning and proceed in this case.

Test Description:
Existing and new unit tests validate the new logic in `blade_class`
property.

Tested the `sat swap blade --action disable` command on baldar, which
was exhibiting the issue where a single blade had some nodes reporting
as "Hill" and some as "Mountain" in HSM. The command worked as expected.

(cherry picked from commit 5cae51386c0c1398c217c3be2fb92d7604a9246d)