CRAYSAT-2000: Fix `EnhancedValidationError.__str__`
Fix the `__str__` method of `EnhancedValidationError` to properly handle
the case where the validation error occurs because the instance object
matches more than exactly one subschema in a `oneOf`. When this occurs,
the `jsonschema.ValidationError` does not contain any `context`, and the
`message` attribute just contains a message to the effect of "<instance>
is valid under each of <oneOf subschemas>".

Fix this issue by explicitly checking whether "is valid under each"
occurs in the message for the `ValidationError` in the `__str__` method.
In addition, if the `context` is empty in the `best_context` method,
simply return the empty list because there is no additional context to
report.

With this fix, it is no longer necessary to add the `not:` subschemas,
so remove them. The schema is more clear without them, but it is clear
now that it was a valid workaround because it resulted in neither of the
subschemas in the `oneOf` matching the instance instead of both
matching.

Test Description:
Tested on three different bootprep files that each define a CFS
configuration with a single source-based layer. When only branch or only
commit is specified, it passes validation. When both are specified it
fails validation with a clear error.

In addition, generated HTML schema documentation with `sat bootprep
generate-docs` and inspected it in the browser.

Finally, also made a temporary modification to `bootprep_schema.yaml`
that added a third property, `tag`, which was mutually exclusive with
the `branch`, and `commit` properties and could be specified instead of
those properties. Verified that an input file that specified two of the
three mutually exclusive properties (e.g. `branch` and `commit`), but
not the third (e.g. `tag`) was still properly caught.