CRAYSAT-2004: Add Jinja2 rendering of `if_exists` property
Add Jinja2 rendering of the `if_exists` property for CFS configurations,
IMS images, and BOS session templates.

It is not sufficient to add the `jinja_rendered` method decorator to the
`if_exists` property.  The bootprep schema file specified an `enum` that
listed the explicit values that could be specified, and the input file
is validated against the schema prior to any values being rendered by
Jinja2.

Address this by removing the `enum` value from the schema and instead
add validation methods for the input items that check the value against
the allowed values after being Jinja-rendered.

Test Description:
Tested by the automated tests implemented in CRAYSAT-2005.