CAST-38539: Warn and prompt in `sat bmccreds` for long passwords
Add logic to `sat bmccreds` to log a warning if the provided or
generated password is longer than 20 characters. This is the maximum
password length supported by `ipmitool`, which is required to control
management nodes during system boot and shutdown procedures.

The warning is logged whenever the password exceeds 20 characters. The
user must answer the prompt if it is determined that the affected xnames
include management node BMCs or if that can't be determined (i.e. if the
HSM query fails, or `--no-hsm-check` is specified).

Added unit tests for the new `validate_password_length` method.

Test Description:
Unit tests pass. Tested the following cases with `sat bmmcreds` on
vinland:

- User-specified password of an acceptable length against a compute and
  management node BMC
- Too long generated password against a compute and management node BMC
- Too long user-specified password against a compute and management node
  BMC
- Too long user-specified password against only compute BMC
- Exactly 20-character password against both BMCs
- Too long password entered at the prompt against both BMCs